[issue-137] Created correct format of components
