+++ cocoon/frontend/src/scheduler/clientScheduler/itineraryDaySelector.js
@@ -1,0 +1,15 @@
+// Import React Components
+import React from 'react'
+import { Component } from 'react';
+
+// Import Cocoon Components
+import ItineraryTimeRangeSelector from "./itineraryTimeRangeSelector";
+
+export default class ItineraryDaySelector extends Component {
+
+    render() {
+        return(
+            <ItineraryTimeRangeSelector/>
+        );
+    }
+}
