+++ cocoon/frontend/src/scheduler/clientScheduler/itineraryTimeSelector.js
@@ -2,42 +2,14 @@
 import React from 'react'
 import { Component } from 'react';
 
-import TimeRange from 'react-time-range';
-import moment from 'moment';
+// Import Cocoon Components
+import ItineraryDaySelector from "./itineraryDaySelector";
 
-class ItineraryTimeSelector extends Component {
-    state = {
-        startTime: moment().startOf('day'),
-        endTime: moment().endOf('day').subtract(30, 'minutes'),
-    };
-
-    returnFunction = (time) => {
-        if (time.endTime) {
-            this.setState({
-                endTime: moment(time.endTime)
-            })
-        }
-        if (time.startTime) {
-            this.setState({
-                startTime: moment(time.startTime)
-            })
-        }
-    };
+export default class ItineraryTimeSelector extends Component {
 
     render() {
-        return (
-            <div>
-                <h2>The Time Selector</h2>
-                <TimeRange
-                    sameIsValid={false}
-                    startMoment={this.state.startTime}
-                    endMoment={this.state.endTime}
-                    onChange={this.returnFunction}
-                />
-            </div>
+        return(
+            <ItineraryDaySelector/>
         );
     }
-
 }
-
-export default ItineraryTimeSelector
